+++ src/rchia/Cats/SetNameCommand.cs
@@ -8,7 +8,7 @@ namespace rchia.Cats;
 internal sealed class SetNameCommand : WalletCommand
 {
     [Option("i", "id", Default = 1, Description = "Id of the CAT wallet to use")]
-    public uint Id { get; init; } = 1;
+    public int Id { get; init; } = 1;
 
     [Option("n", "name", IsRequired = true, Description = "The new name of the wallet")]
     public string Name { get; init; } = string.Empty;
@@ -19,7 +19,7 @@ internal sealed class SetNameCommand : WalletCommand
         return await DoWorkAsync("Setting wallet name...", async output =>
         {
             using var rpcClient = await ClientFactory.Factory.CreateRpcClient(output, this, ServiceNames.Wallet);
-            var wallet = new CATWallet(Id, await Login(rpcClient, output));
+            var wallet = new CATWallet((uint)Id, await Login(rpcClient, output));
             using var cts = new CancellationTokenSource(TimeoutMilliseconds);
 
             await wallet.SetName(Name, cts.Token);
