+++ src/rchia/PlotNft/InspectNftCommand.cs
@@ -8,7 +8,7 @@ namespace rchia.PlotNft;
 internal sealed class InspectNftCommand : WalletCommand
 {
     [Option("i", "id", Default = 1, Description = "Id of the user wallet to use")]
-    public uint Id { get; init; } = 1;
+    public int Id { get; init; } = 1;
 
     [CommandTarget]
     public async Task<int> Run()
@@ -16,7 +16,7 @@ internal sealed class InspectNftCommand : WalletCommand
         return await DoWorkAsync("Retrieving nft plot info...", async output =>
         {
             using var rpcClient = await ClientFactory.Factory.CreateRpcClient(output, this, ServiceNames.Wallet);
-            var wallet = new PoolWallet(Id, await Login(rpcClient, output));
+            var wallet = new PoolWallet((uint)Id, await Login(rpcClient, output));
 
             using var cts = new CancellationTokenSource(TimeoutMilliseconds);
             await wallet.Validate(cts.Token);
